+++ CHANGELOG.md
@@ -9,6 +9,7 @@ Note that since we don't clearly distinguish between a public and private interf
 - Fix ``xrayShader`` & ``ignoreLight`` params not working at the same time
 - Add ``ignoreLight`` to component params
 - Fix representation preset side effects (changing post-processing parameters, see #363)
+- Add Quick Styles panel (reset, illustrative, stylized)
 
 ## [v3.0.2] - 2022-01-30